extract getSuspendedSaleQuery and deleteSuspendedSaleQuery
